+++ types/protobuf.go
@@ -138,8 +138,9 @@ func (pb2tm) ValidatorUpdates(vals []abci.ValidatorUpdate) ([]*Validator, error)
 			return nil, err
 		}
 		updated := NewValidator(pub, v.Power)
-		updated.SetRelayerBlsKey(v.RelayerBlsKey)
+		updated.SetBlsKey(v.BlsKey)
 		updated.SetRelayerAddress(v.RelayerAddress)
+		updated.SetChallengerAddress(v.ChallengerAddress)
 		tmVals[i] = updated
 	}
 	return tmVals, nil