+++ types/validator.go
@@ -13,11 +13,11 @@ import (
 )
 
 const (
-	// RelayerBlsKeySize is the size of validator's relayer bls public key.
-	RelayerBlsKeySize = 48
+	// BlsPubKeySize is the size of validator's relayer bls public key.
+	BlsPubKeySize = 48
 
-	// RelayerAddressSize is the size of validator's relayer address.
-	RelayerAddressSize = 20
+	// AddressSize is the size of validator's relayer address.
+	AddressSize = 20
 )
 
 // Volatile state for each Validator
@@ -30,8 +30,9 @@ type Validator struct {
 
 	ProposerPriority int64 `json:"proposer_priority"`
 
-	RelayerBlsKey  []byte `json:"relayer_bls_key"` // bls public key of authorized relayer/operator
-	RelayerAddress []byte `json:"relayer_address"` // address of authorized relayer/operator
+	BlsKey            []byte `json:"bls_key"`            // bls public key
+	RelayerAddress    []byte `json:"relayer_address"`    // address of authorized relayer/operator
+	ChallengerAddress []byte `json:"challenger_address"` // address of authorized challenger/operator
 }
 
 // NewValidator returns a new validator with the given pubkey and voting power.
@@ -61,11 +62,14 @@ func (v *Validator) ValidateBasic() error {
 		return fmt.Errorf("validator address is the wrong size: %v", v.Address)
 	}
 
-	if len(v.RelayerBlsKey) != 0 && len(v.RelayerBlsKey) != RelayerBlsKeySize {
-		return fmt.Errorf("validator relayer bls key is the wrong size: %v", v.RelayerBlsKey)
+	if len(v.BlsKey) != 0 && len(v.BlsKey) != BlsPubKeySize {
+		return fmt.Errorf("validator relayer bls key is the wrong size: %v", v.BlsKey)
 	}
-	if len(v.RelayerAddress) != 0 && len(v.RelayerAddress) != RelayerAddressSize {
+	if len(v.RelayerAddress) != 0 && len(v.RelayerAddress) != AddressSize {
 		return fmt.Errorf("validator relayer address is the wrong size: %v", v.RelayerAddress)
+	}
+	if len(v.ChallengerAddress) != 0 && len(v.ChallengerAddress) != AddressSize {
+		return fmt.Errorf("validator challenger address is the wrong size: %v", v.ChallengerAddress)
 	}
 
 	return nil
@@ -107,15 +111,19 @@ func (v *Validator) CompareProposerPriority(other *Validator) *Validator {
 // 2. public key
 // 3. voting power
 // 4. proposer priority
+// 5. relayer address
+// 6. challenger address
 func (v *Validator) String() string {
 	if v == nil {
 		return "nil-Validator"
 	}
-	return fmt.Sprintf("Validator{%v %v VP:%v A:%v}",
+	return fmt.Sprintf("Validator{%v %v VP:%v A:%v R:%v C:%v}",
 		v.Address,
 		v.PubKey,
 		v.VotingPower,
-		v.ProposerPriority)
+		v.ProposerPriority,
+		v.RelayerAddress,
+		v.ChallengerAddress)
 }
 
 // ValidatorListString returns a prettified validator list for logging purposes.
@@ -139,10 +147,11 @@ func (v *Validator) Bytes() []byte {
 	}
 
 	pbv := tmproto.SimpleValidator{
-		PubKey:         &pk,
-		VotingPower:    v.VotingPower,
-		RelayerBlsKey:  v.RelayerBlsKey,
-		RelayerAddress: v.RelayerAddress,
+		PubKey:            &pk,
+		VotingPower:       v.VotingPower,
+		BlsKey:            v.BlsKey,
+		RelayerAddress:    v.RelayerAddress,
+		ChallengerAddress: v.ChallengerAddress,
 	}
 
 	bz, err := pbv.Marshal()
@@ -152,14 +161,19 @@ func (v *Validator) Bytes() []byte {
 	return bz
 }
 
-// SetRelayerBlsKey will update the bls public key of relayer.
-func (v *Validator) SetRelayerBlsKey(blsKey []byte) {
-	v.RelayerBlsKey = blsKey
+// SetBlsKey will update the bls public key of relayer.
+func (v *Validator) SetBlsKey(blsKey []byte) {
+	v.BlsKey = blsKey
 }
 
 // SetRelayerAddress will update the relayer address of validator.
 func (v *Validator) SetRelayerAddress(address []byte) {
 	v.RelayerAddress = address
+}
+
+// SetChallengerAddress will update the challenger address of validator.
+func (v *Validator) SetChallengerAddress(address []byte) {
+	v.ChallengerAddress = address
 }
 
 // ToProto converts Validator to protobuf
@@ -174,12 +188,13 @@ func (v *Validator) ToProto() (*tmproto.Validator, error) {
 	}
 
 	vp := tmproto.Validator{
-		Address:          v.Address,
-		PubKey:           pk,
-		VotingPower:      v.VotingPower,
-		ProposerPriority: v.ProposerPriority,
-		RelayerBlsKey:    v.RelayerBlsKey,
-		RelayerAddress:   v.RelayerAddress,
+		Address:           v.Address,
+		PubKey:            pk,
+		VotingPower:       v.VotingPower,
+		ProposerPriority:  v.ProposerPriority,
+		BlsKey:            v.BlsKey,
+		RelayerAddress:    v.RelayerAddress,
+		ChallengerAddress: v.ChallengerAddress,
 	}
 
 	return &vp, nil
@@ -201,8 +216,9 @@ func ValidatorFromProto(vp *tmproto.Validator) (*Validator, error) {
 	v.PubKey = pk
 	v.VotingPower = vp.GetVotingPower()
 	v.ProposerPriority = vp.GetProposerPriority()
-	v.RelayerBlsKey = vp.GetRelayerBlsKey()
+	v.BlsKey = vp.GetBlsKey()
 	v.RelayerAddress = vp.GetRelayerAddress()
+	v.ChallengerAddress = vp.GetChallengerAddress()
 	return v, nil
 }